--- styles/pages/project.module.scss
@@ -1,40 +1,0 @@
-@import "../mixins/variables.mixin.scss";
-@import "../mixins/breakpoints.mixin.scss";
-.cardSection {
-    display: flex;
-    flex-wrap: wrap;
-    width: 90%;
-    margin: auto;
-    text-align: center;
-    justify-content: center;
-}
-
-.heading {
-    width: 85%;
-    margin: 50px auto;
-}
-
-.card {
-    margin: 25px 50px;
-    width: 300px;
-    max-width: 100%;
-    background-color: map-get($map: $colors, $key: seventh);
-    color: map-get($map: $colors, $key: fifth);
-    font-size: 1.25rem;
-    box-shadow: 5px 7px 10px 1px map-get($map: $colors, $key: sixth);
-    transition: all 0.4s ease;
-    img {
-        max-width: 100%;
-        width: 100%;
-        height: calc(100% - 50px);
-        transition: all 0.4s ease;
-    }
-    p {
-        margin: 10px 0;
-        width: 100%;
-    }
-
-    &:hover {
-        transform: scale(1.015);
-    }
-}
